pyproject.toml
[project]
name = "agno-basics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.1.4",
    "openai>=2.3.0",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.44",
]
